Move all pages to separate directory.
