+++ src/courses/f17/cs290/index.js
@@ -1,5 +1,5 @@
 import { generateSitePath } from '../../../lib/SitePath';
-import TAInfoPage from '../../../components/TAInfoPage';
+import TAInfoPage from '../../../pages/TAInfoPage';
 
 import calendar from './calendar';
 import assignments from './assignments';
