+++ src/courses/w18/cs492/FinalProjectProposalPage.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 
 import APIListData from './APIList.json';
-import CourseSubPage from '../../../components/CourseSubPage';
+import CourseSubPage from '../../../pages/CourseSubPage';
 import './AssignmentPage.css';
 
 class FinalProjectProposalPage extends Component {
