+++ src/pages/HomePage.js
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 
 import { currentTerm, courseData } from '../CourseData';
 import { generateSitePath } from '../lib/SitePath';
-import PageContent from './PageContent';
+import PageContent from '../components/PageContent';
 import headshot from '../static/rob.jpg';
 import '../styles/HomePage.css';
