+++ src/pages/LabInfoPage.js
